[Compatibility] working on getting tree generation to work with the new map generation

+++ mc2d/utils.py
@@ -16,7 +16,7 @@ def find_grid_box(x, y):
     left_x = x - (x % int(TILE_SIZE * SCALING))
     bottom_y = y - (y % int(TILE_SIZE * SCALING))
 
-    center_x = left_x + (TILE_SIZE * SCALING) // 2
-    center_y = bottom_y + (TILE_SIZE * SCALING) // 2
+    center_x = left_x + int(TILE_SIZE * SCALING) // 2
+    center_y = bottom_y + int(TILE_SIZE * SCALING) // 2
 
     return center_x, center_y
